Extract message copy fallback
